Moved most of cli.js into a new runner.js
Fixes #283.
